Move VAE and Pertubation to utils/net/continuous.py
